pyproject.toml
[project]
name = "ai-video-editor"
version = "0.1.0"
description = "AI-powered video storyboard generator — analyze trip footage into structured shot-by-shot breakdowns"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "google-genai>=1.0",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]
scene = [
    "scenedetect[opencv]",
]
whisper = [
    "mlx-whisper",
]

[project.scripts]
vx = "ai_video_editor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_video_editor"]

[tool.ruff]
target-version = "py311"
line-length = 100
